homeworks/hw_flask_practice3/requirements.txt
aniso8601==9.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
ddtrace==0.49.3
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mock==4.0.3
packaging==20.9
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
pycparser==2.20
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
pytest-mock==3.6.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.16.0
SQLAlchemy==1.4.18
tenacity==7.0.0
toml==0.10.2
urllib3==1.26.5
Werkzeug==2.0.1
